pyproject.toml
[project]
name = "data-science-team-agent"
version = "1.0.0"
description = "AI Data Science Team is a Python library of specialized agents for common data science workflows, plus a flagship app: AI Pipeline Studio. The Studio turns your work into a visual, reproducible pipeline, while the AI team handles data loading, cleaning, visualization, and modeling."
authors = [{ name = "Paras Chamoli", email = "paras@getbindu.com" }]
readme = "README.md"
keywords = ["python"]
requires-python = ">=3.12,<3.13"

dependencies = [
    "bindu==2026.9.4",
    "pyperclip>=1.8.0",
    "rich>=13.0.0",
    "openai>=2.11.0",
    "sqlalchemy>=1.4.0",
    "mem0ai>=1.0.1",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "plotly>=5.0.0",
    "scikit-learn>=1.3.0",
    "ipython>=8.0.0",
    "psutil>=5.9.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.0.1",
    "requests>=2.31.0",
    "h2o>=3.44.0",
    "mlflow>=2.8.0",
]

classifiers = [
    "Intended Audience :: Developers",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

[project.urls]
Homepage = "https://Paraschamoli.github.io/data-science-team-agent/"
Repository = "https://github.com/Paraschamoli/data-science-team-agent"
Documentation = "https://Paraschamoli.github.io/data-science-team-agent/"

[dependency-groups]
dev = [
    "pytest>=7.2.0",
    "pytest-asyncio>=0.23.0",
    "pre-commit>=2.20.0",
    "tox-uv>=1.11.3",
    "ty>=0.0.1a16",
    "pytest-cov>=4.0.0",
    "ruff>=0.11.5",
    "mkdocs>=1.4.2",
    "mkdocs-material>=8.5.10",
    "mkdocstrings[python]>=0.26.1",
    "mypy>=1.0.0",
    "types-requests>=2.31.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["data_science_team_agent"]

[tool.ty.environment]
python = "./.venv"
python-version = "3.12"

# Correct configuration for ty
[tool.ty.rules]
invalid-argument-type = "ignore"
unresolved-attribute = "ignore"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
target-version = "py312"
line-length = 120
fix = true

[tool.ruff.lint]
select = [
    "YTT",
    "S",
    "B",
    "A",
    "C4",
    "T10",
    "SIM",
    "I",
    "C90",
    "E",
    "W",
    "F",
    "PGH",
    "UP",
    "RUF",
    "TRY",
]

ignore = [
    "E501",
    "E731",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]

[tool.ruff.format]
preview = true

[tool.coverage.report]
skip_empty = true

[tool.coverage.run]
branch = true
source = ["data_science_team_agent"]

[tool.mypy]
python_version = "3.12"
warn_return_any = false
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = false
disallow_untyped_decorators = false
no_implicit_optional = false
warn_redundant_casts = false
warn_unused_ignores = false
warn_no_return = false
warn_unreachable = false
strict_equality = false

[[tool.mypy.overrides]]
module = [
    "tests.*",
    "docs.*",
]
ignore_errors = true

[[tool.mypy.overrides]]
module = [
    "langchain_core.*",
    "langgraph.*",
    "bindu.*",
    "mem0ai.*",
    "pandas.*",
    "numpy.*",
    "plotly.*",
    "scikit-learn.*",
    "ipython.*",
    "psutil.*",
    "httpx.*",
    "python_dotenv.*",
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
    "data_science_team_agent.*",
]
allow_any_explicit = true
allow_any_unimported = true
